+++ docker-images/push.sh
@@ -5,9 +5,6 @@ TAG="$1"
 docker push walmartlabs/concord-server:${TAG}
 docker push walmartlabs/concord-server:latest
 
-docker push walmartlabs/concord-console:${TAG}
-docker push walmartlabs/concord-console:latest
-
 docker push walmartlabs/concord-ansible:${TAG}
 docker push walmartlabs/concord-ansible:latest