+++ server/impl/src/main/java/com/walmartlabs/concord/server/cfg/ServerConfiguration.java
@@ -20,6 +20,8 @@ package com.walmartlabs.concord.server.cfg;
  * =====
  */
 
+import org.eclipse.jetty.server.CustomRequestLog;
+
 import java.io.Serializable;
 
 /**
@@ -29,30 +31,29 @@ import java.io.Serializable;
  */
 public class ServerConfiguration implements Serializable {
 
+    public static final String ACCESS_LOG_FORMAT = CustomRequestLog.EXTENDED_NCSA_FORMAT + " %{ms}T";
+
+    private static final String ACCESS_LOG_PATH_KEY = "ACCESS_LOG_PATH";
+    private static final String ACCESS_LOG_RETAIN_DAYS_KEY = "ACCESS_LOG_RETAIN_DAYS";
     private static final String API_PORT_KEY = "API_PORT";
+    private static final String BASE_RESOURCE_PATH_KEY = "BASE_RESOURCE_PATH";
     private static final String SECURE_COOKIES_KEY = "SECURE_COOKIES";
     private static final String SESSION_TIMEOUT_KEY = "SESSION_TIMEOUT";
 
     public static final int port;
     public static final boolean secureCookies;
     public static final int sessionTimeout;
+    public static final String accessLogPath;
+    public static final int accessLogRetainDays;
+    public static final String baseResourcePath;
 
     static {
         port = Integer.parseInt(getEnv(API_PORT_KEY, "8001"));
         secureCookies = Boolean.parseBoolean(getEnv(SECURE_COOKIES_KEY, "false"));
         sessionTimeout = Integer.parseInt(getEnv(SESSION_TIMEOUT_KEY, "1800")); // 30 min
-    }
-
-    public int getPort() {
-        return port;
-    }
-
-    public boolean isSecureCookies() {
-        return secureCookies;
-    }
-
-    public int getSessionTimeout() {
-        return sessionTimeout;
+        accessLogPath = getEnv(ACCESS_LOG_PATH_KEY, null);
+        accessLogRetainDays = Integer.parseInt(getEnv(ACCESS_LOG_RETAIN_DAYS_KEY, "7"));
+        baseResourcePath = getEnv(BASE_RESOURCE_PATH_KEY, null);
     }
 
     private static String getEnv(String k, String defaultValue) {